requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cloudinary==1.42.2
colorama==0.4.6
Django==4.2.20
django-browser-reload==1.18.0
geojson==3.2.0
gunicorn==23.0.0
packaging==24.2
pillow==10.4.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
python-dotenv==1.0.1
qrcode==8.1
rsa==4.9
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
xlwt==1.3.0
